refactor aiguard files
