+++ ddtrace/aiguard/_streaming.py
@@ -23,7 +23,7 @@ import wrapt
 
 from ddtrace.aiguard._context import is_aiguard_context_active
 import ddtrace.internal.logger as ddlogger
-from ddtrace.internal.settings.asm import ai_guard_config
+from ddtrace.internal.settings.aiguard import aiguard_config
 from ddtrace.llmobs._integrations.base_stream_handler import AsyncStreamHandler
 from ddtrace.llmobs._integrations.base_stream_handler import BaseStreamHandler
 
@@ -115,7 +115,7 @@ class BufferedAIGuardStream(wrapt.ObjectProxy):  # type: ignore[misc]  # wrapt s
         if self._self_passthrough:
             return None
         if self._self_chunks is None:
-            if not ai_guard_config._ai_guard_analyze_stream_responses_enabled or is_aiguard_context_active():
+            if not aiguard_config._ai_guard_analyze_stream_responses_enabled or is_aiguard_context_active():
                 self._self_passthrough = True
                 return None
             chunks = list(self.__wrapped__)  # drives contrib tracing + finalize_stream
@@ -226,7 +226,7 @@ class BufferedAIGuardAsyncStream(wrapt.ObjectProxy):  # type: ignore[misc]  # wr
         if self._self_passthrough:
             return None
         if self._self_chunks is None:
-            if not ai_guard_config._ai_guard_analyze_stream_responses_enabled or is_aiguard_context_active():
+            if not aiguard_config._ai_guard_analyze_stream_responses_enabled or is_aiguard_context_active():
                 self._self_passthrough = True
                 return None
             chunks: list[Any] = []
