+++ ddtrace/aiguard/_trace_utils.py
@@ -1,0 +1,16 @@
+from typing import TYPE_CHECKING
+
+from ddtrace.constants import USER_KEEP
+
+
+if TYPE_CHECKING:
+    from ddtrace._trace.span import Span
+
+
+def _aiguard_manual_keep(span: "Span") -> None:
+    from ddtrace.internal.constants import SAMPLING_DECISION_TRACE_TAG_KEY
+    from ddtrace.internal.sampling import SamplingMechanism
+
+    span._override_sampling_decision(USER_KEEP)
+    # set decision maker to AI_GUARD = -13
+    span._set_attribute(SAMPLING_DECISION_TRACE_TAG_KEY, f"-{SamplingMechanism.AI_GUARD}")
