+++ ddtrace/aiguard/integrations/_openai_chat.py
@@ -18,7 +18,7 @@ from ddtrace.aiguard._common import _get
 from ddtrace.aiguard._context import is_aiguard_context_active
 from ddtrace.aiguard.integrations._openai import _wrap_abort_error
 import ddtrace.internal.logger as ddlogger
-from ddtrace.internal.settings.asm import ai_guard_config
+from ddtrace.internal.settings.aiguard import aiguard_config
 
 
 logger = ddlogger.get_logger(__name__)
@@ -179,7 +179,7 @@ def _openai_chat_completion_before(client: AIGuardClient, kwargs: dict[str, Any]
 
     logger.debug("AI Guard openai before-hook evaluating %d message(s)", len(ai_guard_messages))
     try:
-        client.evaluate(ai_guard_messages, Options(block=ai_guard_config._ai_guard_block))
+        client.evaluate(ai_guard_messages, Options(block=aiguard_config._ai_guard_block))
     except AIGuardAbortError as e:
         raise _wrap_abort_error(e)
     except Exception:
@@ -212,7 +212,7 @@ def _openai_chat_completion_after(client: AIGuardClient, kwargs: dict[str, Any],
     all_messages = request_messages + response_messages
 
     try:
-        client.evaluate(all_messages, Options(block=ai_guard_config._ai_guard_block))
+        client.evaluate(all_messages, Options(block=aiguard_config._ai_guard_block))
     except AIGuardAbortError as e:
         raise _wrap_abort_error(e)
     except Exception:
