+++ ddtrace/appsec/_constants.py
@@ -5,15 +5,13 @@ from _io import StringIO
 import os
 from re import Match
 import sys
-from typing import Any
-from typing import Generator
-from typing import Iterator
 from typing import Literal  # noqa:F401
 
 from ddtrace.internal.constants import HTTP_REQUEST_BLOCKED
 from ddtrace.internal.constants import REQUEST_PATH_PARAMS
 from ddtrace.internal.constants import RESPONSE_HEADERS
 from ddtrace.internal.constants import STATUS_403_TYPE_AUTO
+from ddtrace.internal.constants import Constant_Class  # noqa:F401
 
 
 TEXT_TYPES = (str, bytes, bytearray)
@@ -25,36 +23,6 @@ if sys.version_info >= (3, 14):
     from string.templatelib import Template as TemplateType
 
     TAINTEABLE_TYPES += (TemplateType,)
-
-
-class Constant_Class(type):
-    """
-    metaclass for Constant Classes
-    - You can access constants with APPSEC.ENV or APPSEC["ENV"]
-    - Direct assignment will fail: APPSEC.ENV = "something" raise TypeError, like other immutable types
-    - Constant Classes can be iterated:
-        for constant_name, constant_value in APPSEC: ...
-    """
-
-    def __setattr__(self, __name: str, __value: Any) -> None:
-        raise TypeError("Constant class does not support item assignment: %s.%s" % (self.__name__, __name))
-
-    def __iter__(self) -> Iterator[tuple[str, Any]]:
-        def aux() -> Generator[tuple[str, Any], Any, None]:
-            for t in self.__dict__.items():
-                if not t[0].startswith("_"):
-                    yield t
-
-        return aux()
-
-    def get(self, k: str, default: Any = None) -> Any:
-        return self.__dict__.get(k, default)
-
-    def __contains__(self, k: str) -> bool:
-        return k in self.__dict__
-
-    def __getitem__(self, k: str) -> Any:
-        return self.__dict__[k]
 
 
 class APPSEC(metaclass=Constant_Class):
