+++ ddtrace/appsec/_trace_utils.py
@@ -54,15 +54,6 @@ def _asm_manual_keep(span: Span) -> None:
     span.context._meta[APPSEC.PROPAGATION_HEADER] = "02"
 
 
-def _aiguard_manual_keep(span: Span) -> None:
-    from ddtrace.internal.constants import SAMPLING_DECISION_TRACE_TAG_KEY
-    from ddtrace.internal.sampling import SamplingMechanism
-
-    span._override_sampling_decision(USER_KEEP)
-    # set decision maker to AI_GUARD = -13
-    span._set_attribute(SAMPLING_DECISION_TRACE_TAG_KEY, f"-{SamplingMechanism.AI_GUARD}")
-
-
 def _handle_metadata(entry_span: Span, prefix: str, metadata: dict) -> None:
     MAX_DEPTH = 6
     stack = [(prefix, metadata, 1)]
