+++ pyproject.toml
@@ -86,6 +86,7 @@ datadog = "ddtrace.contrib.internal.mlflow.auth_plugin:DatadogHeaderProvider"
 "remote-configuration" = "ddtrace.internal.remoteconfig.products.client"
 "symbol-database" = "ddtrace.internal.symbol_db.product"
 "appsec" = "ddtrace.internal.appsec.product"
+"aiguard" = "ddtrace.internal.aiguard.product"
 "iast" = "ddtrace.internal.iast.product"
 "sca" = "ddtrace.internal.sca.product"
 "tracer" = "ddtrace._trace.product"
